moved function code in messaging to controller file
